Add files via upload
This file contains VBA code attached to a worksheet to automate the process of stock taking for a restaurant. The process is such that the dailyStockSheet worksheet has a button upon which, on the click event carries out the end of day transaction by summarizing the day's sales, generates a sales pivot table, updates the stock levels while;
1. highlighting the items below stock level and 
2. Populating the opening stock for the following business day.
Finally the procedure advances the business day to the next in readiness for opening business.

Feel free to make inquiry / comment about the code.